LibC: Move bzero() and bcopy() per Dr. POSIX
