+++ DevTools/UserspaceEmulator/Emulator.cpp
@@ -45,6 +45,7 @@
 #include <serenity.h>
 #include <stdio.h>
 #include <string.h>
+#include <strings.h>
 #include <sys/ioctl.h>
 #include <sys/mman.h>
 #include <sys/select.h>
