+++ Libraries/LibC/string.cpp
@@ -38,16 +38,6 @@
 
 extern "C" {
 
-void bzero(void* dest, size_t n)
-{
-    memset(dest, 0, n);
-}
-
-void bcopy(const void* src, void* dest, size_t n)
-{
-    memmove(dest, src, n);
-}
-
 size_t strspn(const char* s, const char* accept)
 {
     const char* p = s;
