+++ Libraries/LibC/strings.h
@@ -33,5 +33,7 @@ __BEGIN_DECLS
 
 int strcasecmp(const char*, const char*);
 int strncasecmp(const char*, const char*, size_t);
+void bzero(void*, size_t);
+void bcopy(const void*, void*, size_t);
 
 __END_DECLS
