+++ Services/AudioServer/Mixer.cpp
@@ -30,6 +30,7 @@
 #include <AudioServer/ClientConnection.h>
 #include <AudioServer/Mixer.h>
 #include <pthread.h>
+#include <strings.h>
 
 namespace AudioServer {
